Add logos for Eth and Sol

+++ app/src/main/java/com/initial/ventum/network/CoinData.kt
@@ -6,5 +6,5 @@ data class CoinData(
     // map the "name" value from returned JSON to "name" in the class
     @Json(name = "name") val name: String,
     //  map the "current_price" value from returned JSON to "currentPrice" in the class
-    @Json(name = "current_price") val currentPrice: Int
+    @Json(name = "current_price") val currentPrice: Float
 )
